[refactor/TK-202]: 스케줄러와 비즈니스 로직 분리
