+++ src/main/java/withbeetravel/service/notification/SettlementRequestLogServiceIml.java
@@ -7,7 +7,9 @@ import withbeetravel.dto.request.settlementRequestLog.SettlementRequestLogDto;
 import withbeetravel.repository.SettlementRequestLogRepository;
 import withbeetravel.repository.SettlementRequestRepository;
 import withbeetravel.repository.TravelMemberRepository;
+import withbeetravel.repository.TravelRepository;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -17,6 +19,9 @@ public class SettlementRequestLogServiceIml implements SettlementRequestLogServi
 
     private final SettlementRequestLogRepository settlementRequestLogRepository;
     private final SettlementRequestRepository settlementRequestRepository;
+    private final TravelRepository travelRepository;
+    private final TravelMemberRepository travelMemberRepository;
+
 
     @Override
     public List<SettlementRequestLogDto> getSettlementRequestLogs(Long userId) {
@@ -33,6 +38,29 @@ public class SettlementRequestLogServiceIml implements SettlementRequestLogServi
         }
 
         return settlementRequestLogDtos;
+    }
+
+    @Override
+    public void createSettlementLogsForEndedTravels() {
+        List<Travel> travels = travelRepository.findAllByTravelEndDate(LocalDate.now());
+        for (Travel travel : travels) {
+            createSettlementRequestLog(travel);
+        }
+    }
+
+    private void createSettlementRequestLog(Travel travel) {
+        List<TravelMember> travelMembers = travelMemberRepository.findAllByTravelId(travel.getId());
+        travelMembers.forEach(travelMember -> {
+            User user = travelMember.getUser();
+            SettlementRequestLog settlementRequestLog = SettlementRequestLog.builder()
+                    .travel(travel)
+                    .user(user)
+                    .logTitle(LogTitle.PAYMENT_REQUEST)
+                    .logMessage(LogTitle.PAYMENT_REQUEST.getMessage(travel.getTravelName()))
+                    .build();
+            settlementRequestLogRepository.save(settlementRequestLog);
+        });
+
     }
 
     private SettlementRequestLogDto createSettlementRequestLogDto(SettlementRequestLog settlementRequestLog, String link) {
